gDRstyle: A package with style requirements for the gDR suite

Package fills a helper package role for whole gDR suite. It helps to support good development practices by keeping style requirements and style tests for other packages. It also contains build helpers to make all package requirements met.

Value

package help page

Note

To learn more about functions start with help(package = "gDRstyle")

See also

Author

Maintainer: Arkadiusz Gladki gladki.arkadiusz@gmail.com (ORCID)

Authors:

Allison Vuong

Dariusz Scigocki

Marcin Kamianowski

Janina Smola

Bartosz Czech (ORCID)

Other contributors:

Aleksander Chlebowski [contributor]